+++ lib/src/export.dart
@@ -74,6 +74,7 @@ export './utils/data_util.dart';
 export './heatmap/heatmap_chart.dart';
 export './heatmap/heatmap_data.dart';
 export './heatmap/seasonality_chart.dart';
+export './bullet/bullet_chart.dart';
 export './pairs/pair_spread_chart.dart';
 export './pie/pie_chart.dart';
 export './radar/radar_chart.dart';